work/requirements.txt
Django
djangorestframework
dj-database-url
psycopg2-binary
factory_boy
pytest
pytest-django

